Relax solidus_auth_devise dependency
This fixes Missing helper file helpers/spree/admin/tables_helper.rb
which happens when using a new solidus with an old solidus_auth_devise.

Technically this is a problem with solidus_auth_devise 1.6.4 which does
not explicitly require an old solidus but I think it's more reasonable
to just let the installer of this gem sort out their own deps.