Change add property CSS selector
We changed the HTML element to add new table elements in
Solidus core from a to button and this is causing this
feature to stop working since it's not more able to find
the link to add the row programmatically.

This new version does not depend on the link being a specific
HTML element, so it works with all Solidus versions.